[#549] Refactor data module packages
