+++ template-compose/app/src/main/java/co/nimblehq/template/compose/di/modules/MoshiModule.kt
@@ -1,6 +1,6 @@
 package co.nimblehq.template.compose.di.modules
 
-import co.nimblehq.template.compose.data.services.providers.MoshiBuilderProvider
+import co.nimblehq.template.compose.data.remote.providers.MoshiBuilderProvider
 import com.squareup.moshi.Moshi
 import dagger.Module
 import dagger.Provides
