+++ template-compose/app/src/main/java/co/nimblehq/template/compose/di/modules/RepositoryModule.kt
@@ -1,7 +1,7 @@
 package co.nimblehq.template.compose.di.modules
 
+import co.nimblehq.template.compose.data.remote.services.ApiService
 import co.nimblehq.template.compose.data.repositories.RepositoryImpl
-import co.nimblehq.template.compose.data.services.ApiService
 import co.nimblehq.template.compose.domain.repositories.Repository
 import dagger.Module
 import dagger.Provides
