+++ template-compose/data/src/main/java/co/nimblehq/template/compose/data/local/storages/BaseSharedPreferences.kt
@@ -1,4 +1,4 @@
-package co.nimblehq.template.compose.data.storages
+package co.nimblehq.template.compose.data.local.storages
 
 import android.content.SharedPreferences
