+++ template-compose/data/src/main/java/co/nimblehq/template/compose/data/remote/providers/ApiServiceProvider.kt
@@ -1,6 +1,6 @@
-package co.nimblehq.template.compose.data.services.providers
+package co.nimblehq.template.compose.data.remote.providers
 
-import co.nimblehq.template.compose.data.services.ApiService
+import co.nimblehq.template.compose.data.remote.services.ApiService
 import retrofit2.Retrofit
 
 object ApiServiceProvider {
