+++ template-compose/data/src/main/java/co/nimblehq/template/compose/data/repositories/RepositoryImpl.kt
@@ -1,8 +1,8 @@
 package co.nimblehq.template.compose.data.repositories
 
 import co.nimblehq.template.compose.data.extensions.flowTransform
-import co.nimblehq.template.compose.data.responses.toModels
-import co.nimblehq.template.compose.data.services.ApiService
+import co.nimblehq.template.compose.data.remote.models.responses.toModels
+import co.nimblehq.template.compose.data.remote.services.ApiService
 import co.nimblehq.template.compose.domain.models.Model
 import co.nimblehq.template.compose.domain.repositories.Repository
 import kotlinx.coroutines.flow.Flow
